made many tests and made all of the files norminette appoiriate.
