+++ and_or_parser/parser_and_or.c
@@ -6,25 +6,11 @@
 /*   By: kuzyilma <kuzyilma@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/20 19:15:24 by kuzyilma          #+#    #+#             */
-/*   Updated: 2025/06/02 14:22:35 by kuzyilma         ###   ########.fr       */
+/*   Updated: 2025/06/02 15:11:00 by kuzyilma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../main/minishell.h"
-
-void print_result(t_split *test)
-{
-    int i;
-    printf("\n\tsize: %d", test->size);
-    printf("\n\tstart:%3p", test->start);
-    i = 0;
-    while(i <= test->size)
-    {
-        printf("\n\t\t%-2d:%s", i, test->start[i]);
-        i++;
-    }
-    printf(" -> after last value\n");
-}
 
 // Cuts out the outermost (). New string == empty? Remove it from 'split'
 static void	cut_out_par(t_split *split)
@@ -54,88 +40,68 @@ static void	cut_out_par(t_split *split)
 		split->size--;
 }
 
-// counts exact str in split's str array . if flag = 1 does not count str in paranthesis
-int	count_str_split(t_split split, const char *str, int flag)
+char	*get_cut_indexs(t_split split)
 {
-	int i;
-	int count;
-	int par;
+	int		i;
+	int		par;
+	char	*ret;
+	int		check;
 
-	i = 0;
-	count = 0;
-	par = 0;
-	while (i < split.size)
-	{
-		par += countchr_not_quote(split.start[i], '(');
-		if ((flag == 0 || par == 0) && (strncmp(str, split.start[i], 4) == 0))
-			count++;
-		par -= countchr_not_quote(split.start[i], ')');
-		i++;
-	}
-	return (count);
-}
-
-char *get_cut_indexs(t_split split)
-{
-	int i;
-	int par;
-	char *ret;
-	int check;
-
-	i = 0;
+	i = -1;
 	par = 0;
 	check = 0;
-	ret = ft_calloc(count_str_split(split, "||", 1) + count_str_split(split, "&&", 1) + 1, sizeof(char));
+	ret = ft_calloc(count_str_split(split, "||", 1) + \
+	count_str_split(split, "&&", 1) + 1, sizeof(char));
 	if (ret == NULL)
 		return (NULL);
-	while (i < split.size)
+	while (++i < split.size)
 	{
 		par += countchr_not_quote(split.start[i], '(');
-		if (par == 0) 
+		if (par == 0)
 		{
-			if(strncmp("||", split.start[i], 3) == 0)
+			if (strncmp("||", split.start[i], 3) == 0)
 				ret[check++] = '1';
 			else if (strncmp("&&", split.start[i], 3) == 0)
 				ret[check++] = '0';
 		}
 		par -= countchr_not_quote(split.start[i], ')');
-		i++;
 	}
 	return (ret);
 }
 
-void parse_and_or(t_shell *shell, t_split split, char *cut_indexs)
+void	parse_and_or(t_shell *shell, t_split split, char *c_i)
 {
-	int i;
-	int par;
-	int start;
-	int check;
+	int	i;
+	int	par;
+	int	st;
+	int	check;
 
 	i = -1;
 	par = 0;
-	start = 0;
+	st = 0;
 	check = -2;
 	while (++i <= split.size)
 	{
 		par += countchr_not_quote(split.start[i], '(');
-		if ((par == 0) && ((i == split.size) || (strncmp("&&", split.start[i], 4) == 0) || (strncmp("||", split.start[i], 4) == 0)))
+		if ((par == 0) && ((i == split.size) || (strncmp("&&", split.start[i], \
+		4) == 0) || (strncmp("||", split.start[i], 4) == 0)))
 		{
 			if (split.start[i] != NULL)
 				free(split.start[i]);
 			split.start[i] = NULL;
-			if (++check == -1 || \
-				(cut_indexs[check] == '0' && shell->past_exit_status == 0) || \
-				(cut_indexs[check] == '1' && shell->past_exit_status != 0))
-				parser_and_or(shell, create_split(&(split.start[start]), i - start));
-			start = i + 1;
+			if (++check == -1 || (c_i[check] == '0' && shell->past_exit_status \
+			== 0) || (c_i[check] == '1' && shell->past_exit_status != 0))
+				parser_and_or(shell, create_split(&(split.start[st]), i - st));
+			st = i + 1;
 		}
 		par -= countchr_not_quote(split.start[i], ')');
 	}
 }
-int paranthesis_parity_check(t_split split)
+
+int	paranthesis_parity_check(t_split split)
 {
-	int i;
-	int par;
+	int	i;
+	int	par;
 
 	i = 0;
 	par = 0;
@@ -149,6 +115,7 @@ int paranthesis_parity_check(t_split split)
 		return (1);
 	return (0);
 }
+
 // Parser for && and || with () priority
 void	parser_and_or(t_shell *shell, t_split split)
 {
@@ -169,10 +136,10 @@ void	parser_and_or(t_shell *shell, t_split split)
 	}
 	if (count_str_split(split, "||", 1) + count_str_split(split, "&&", 1) > 0)
 	{
-	cut_indexs = get_cut_indexs(split);
-	parse_and_or(shell, split, cut_indexs);
-	free(cut_indexs);
+		cut_indexs = get_cut_indexs(split);
+		parse_and_or(shell, split, cut_indexs);
+		free(cut_indexs);
 	}
 	else
-		{printf("%s\n", split.start[0]);}
+		{/*executer*/}
 }
