+++ main/minishell.h
@@ -6,7 +6,7 @@
 /*   By: kuzyilma <kuzyilma@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/08 10:53:30 by kuzyilma          #+#    #+#             */
-/*   Updated: 2025/04/15 14:03:44 by kuzyilma         ###   ########.fr       */
+/*   Updated: 2025/06/02 15:08:43 by kuzyilma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,5 +41,6 @@ int		countchr_not_quote(char*str, char c);
 int		check_single_par(t_split split);
 int		check_symbol(t_split split, char *find, int flag);
 void	sep_opt_arg(t_shell *shell);
+int		count_str_split(t_split split, const char *str, int flag);
 
 #endif
